Rename DataPutter to TestMaker, method makeTest (addData) gets only index
